Develop

+++ sharespace/client/controllers/posts.js
@@ -45,13 +45,29 @@ myApp.controller('PostsController', ['$scope', '$http', '$location', '$routePara
 	// Scope function that adds a post
 	$scope.addPost = function(){
 
-		// Convert string of tags into list of tags
-		$scope.post.tag_list = $scope.post.tag_list.split(" ");
+		// If tags exist
+		if ($scope.post.tag_list) {
 
-		// Remove all duplicates from list of tags
-		$scope.post.tag_list = $scope.post.tag_list.filter( function(item, index, inputArray) {
-           return inputArray.indexOf(item) == index;
-        });
+			// Convert string of tags into list of tags
+			$scope.post.tag_list = $scope.post.tag_list.split(" ");
+
+			// For every element in the array
+			for(var i = $scope.post.tag_list.length - 1; i >= 0; i--) {
+
+				// If there are any extra spaces
+				if($scope.post.tag_list[i] === "") {
+
+					// Remove them from the array
+					$scope.post.tag_list.splice(i, 1);
+
+				}
+			}
+
+			// Remove all duplicates from list of tags
+			$scope.post.tag_list = $scope.post.tag_list.filter( function(item, index, inputArray) {
+				return inputArray.indexOf(item) == index;
+        	});
+		}
 
 		// Define image paths
 		$scope.post.image_paths = [];
@@ -139,6 +155,18 @@ myApp.controller('PostsController', ['$scope', '$http', '$location', '$routePara
 		// Convert tag string of tags into list of tags
 		$scope.post.tag_list = $scope.post.tag_list.split(" ");
 
+		// For every element in the array
+		for(var i = $scope.post.tag_list.length - 1; i >= 0; i--) {
+
+			// If there are any extra spaces
+			if($scope.post.tag_list[i] === "") {
+
+				// Remove them from the array
+				$scope.post.tag_list.splice(i, 1);
+
+			}
+		}
+
 		// Remove all duplicates from list of tags
 		$scope.post.tag_list = $scope.post.tag_list.filter( function(item, index, inputArray) {
            return inputArray.indexOf(item) == index;
@@ -156,19 +184,31 @@ myApp.controller('PostsController', ['$scope', '$http', '$location', '$routePara
 	// We are not getting id from url. This time we are getting it from parameter.
 	$scope.removePost = function(id, image_paths, file_paths){
 
-		// Concatenate paths of images and files in one array
-		var paths = image_paths.concat(file_paths);
+		// Create dialog box that user confirms they want to delete post
+		var result = confirm("Are you sure you want to delete this post?");
 
-		// PUT request that deletes images of the above paths
-		$http.put('/api/multer', paths).success(function(response){
+		// Case where they want to delete post
+		if (result) {
 
-			// DELETE request to delete a post
-			$http.delete('/api/posts/' + id).success(function(response){
+			// Concatenate paths of images and files in one array
+			var paths = image_paths.concat(file_paths);
 
-				// Redirect
-				window.location.href='#/posts';
+			// PUT request that deletes images of the above paths
+			$http.put('/api/multer', paths).success(function(response){
 
+				// DELETE request to delete a post
+				$http.delete('/api/posts/' + id).success(function(response){
+
+					// Redirect
+					window.location.href='#/posts';
+				});
 			});
-		});
+
+		// Case where they don't want to delete post
+		} else {
+
+			// Redirect to post
+			window.location.href='#/posts/details/' + id;
+		}
 	}
 }]);
